requirements.txt
azure-core==1.35.0
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
idna==3.10
isodate==0.7.2
pycparser==2.22
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
